Work on schema2 SQL like variants' query interface.
